Implement InvokeDynamic for getMember
BaseObject#getMember is now much speedier due to call site
inline caching! Currently, the cache is simply monomorphic,
and although it's likely we'd see further performance improvements
if this was changed to be polymorphic, that can come later.

Speaking of performance, my testing (with mandelbrot.psc) saw
profile times of the getMember call drop from 9,200ms to 6,580ms.
That's a performance gain of about 30%!